+++ backend/index.ts
@@ -62,7 +62,7 @@ const service = {
 
 				const logger = new LogRecorder({enabled: log});
 
-				const pages = result.svgs.map(svg => staffSvg.parseSvgPage(svg, source, {DOMParser, logger, attributes}));
+				const pages = result.svgs.map(svg => staffSvg.parseSvgPage(svg, source, lilyDocument, {DOMParser, logger, attributes}));
 				const doc = {
 					__prototype: "SheetDocument",
 					pages: pages.map(page => page.structure),