+++ inc/staffSvg/index.ts
@@ -9,10 +9,28 @@ import * as StaffNotation from "./staffNotation";
 import TextSource from "../textSource";
 // eslint-disable-next-line
 import LilyDocument from "../lilyParser/lilyDocument";
+// eslint-disable-next-line
+import {LilyDocumentAttributeReadOnly} from "../lilyParser/lilyDocument";
 
 
 
-const parseSvgPage = (dom, source: string | TextSource, lilyDocument: LilyDocument, {logger = new LogRecorder(), attributes, ...options}) => {
+interface SvgPageParserOptions {
+	lilyDocument?: LilyDocument;
+	logger?: LogRecorder;
+	attributes?: LilyDocumentAttributeReadOnly;
+	tieLocations?: {[key: string]: boolean};
+
+	DOMParser?: any;
+};
+
+
+const parseSvgPage = (dom, source: string | TextSource, {
+	lilyDocument,
+	logger = new LogRecorder(),
+	attributes,
+	tieLocations,
+	...options
+}: SvgPageParserOptions = {}) => {
 	if (!(source instanceof TextSource))
 		source = new TextSource(source);
 
@@ -23,7 +41,7 @@ const parseSvgPage = (dom, source: string | TextSource, lilyDocument: LilyDocume
 		return {structure: null, hashTable: {}};
 
 	if (!attributes && lilyDocument)
-		attributes = lilyDocument.globalAttributes({readonly: true});
+		attributes = lilyDocument.globalAttributes({readonly: true}) as LilyDocumentAttributeReadOnly;
 
 	const {tokens, hashTable} = tokenizeElements(elem.children, attributes, logger);
 
@@ -31,14 +49,15 @@ const parseSvgPage = (dom, source: string | TextSource, lilyDocument: LilyDocume
 	const viewBox = {x, y, width, height};
 
 	// mark tie symbol on tokens
-	const tieLocations = lilyDocument.getTiedNoteLocations(source).reduce((table, loc) => ((table[`${loc[0]}:${loc[1]}`] = true), table), {});
-	tokens.forEach(token => {
-		if (token.sourcePosition) {
-			const {line, start} = token.sourcePosition;
-			if (tieLocations[`${line}:${start}`])
-				token.addSymbol("TIE");
-		}
-	});
+	if (tieLocations) {
+		tokens.forEach(token => {
+			if (token.sourcePosition) {
+				const {line, start} = token.sourcePosition;
+				if (tieLocations[`${line}:${start}`])
+					token.addSymbol("TIE");
+			}
+		});
+	}
 
 	return {
 		structure: organizeTokens(tokens, source, {logger, viewBox, width: elem.width, height: elem.height}),
@@ -51,11 +70,13 @@ const createSheetDocumentFromSvgs = (svgs: string[], ly: string, lilyDocument: L
 	doc: SheetDocument,
 	hashTable: {[key: string]: object},
 } => {
-	const attributes = lilyDocument.globalAttributes({readonly: true});
+	const attributes = lilyDocument.globalAttributes({readonly: true}) as LilyDocumentAttributeReadOnly;
 
 	const source = new TextSource(ly);
 
-	const pages = svgs.map(svg => parseSvgPage(svg, source, lilyDocument, {DOMParser, logger, attributes}));
+	const tieLocations = lilyDocument.getTiedNoteLocations(source).reduce((table, loc) => ((table[`${loc[0]}:${loc[1]}`] = true), table), {});
+
+	const pages = svgs.map(svg => parseSvgPage(svg, source, {DOMParser, logger, attributes, tieLocations}));
 	const doc = new SheetDocument({
 		pages: pages.map(page => page.structure),
 	});