+++ inc/staffSvg/staffNotation.ts
@@ -406,7 +406,8 @@ const parseNotationInMeasure = (context: StaffContext, measure) => {
 		if (token.is("ALTER")) {
 			// ignore invalid alters
 			if (Number.isInteger(token.ry * 2)) {
-				if (token.logicX < measure.headX || (token.source && token.source.substr(0, 4) === "\\key"))
+				//if (token.logicX < measure.headX || (token.source && token.source.substr(0, 4) === "\\key"))
+				if (token.is("KEY") || token.logicX < measure.headX)
 					context.setKeyAlter(token.ry, token.alterValue);
 				else
 					context.setAlter(token.ry, token.alterValue);