+++ changelog.md
@@ -1,10 +1,12 @@
 
 ## TODO
+	batch MIDI match checking
 	measures clipping, playground.sliceSource
 	measure notes in Notation parsing
 	music duration parsing
 		parse measure ticks from ly
 	staffSvg
+		treble_8
 		system splitting, chord mode symbol affiliation
 		staffNotation.ts: clusterize factor issue
 		remove empty track	(6.1 落叶.ly)